% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mainfunction.R
\name{DataInfo}
\alias{DataInfo}
\title{Exhibit basic data information}
\usage{
DataInfo(data, diversity = "TD", datatype = "abundance", tree = NULL)
}
\arguments{
\item{data}{\code{data.frames}}

\item{diversity}{selection of diversity type: \code{'TD'} = 'Taxonomic diversity', \code{'PD'} = 'Phylogenetic diversity', and \code{'FD'} = 'Functional diversity'.}

\item{datatype}{data type of input data: individual-based abundance data \code{(datatype = "abundance")}, sampling-unit-based incidence frequencies data \code{(datatype = "incidence")},}
}
\value{
a data.frame of basic data information incliuding sample size, observed species richness, sample coverage estimate, and the first ten abundance frequency counts.
}
\description{
\code{DataInfo}: exhibits basic data information
}
\examples{
## Taxonomic diversity
data(macro)
DataInfo(data = macro, diversity = 'TD')


## Phylogenetic diversity
data(macro)
data(macro_tree)
DataInfo(data = macro, diversity = 'PD', tree = macro_tree)


## Functional diversity
data(macro)
DataInfo(data = macro, diversity = 'FD')
}
