% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mainfunction.R
\name{gghier_functional}
\alias{gghier_functional}
\title{ggplot2 extension for outcome from \code{hier_functional}}
\usage{
gghier_functional(outcome, method = 1, profile = "q")
}
\arguments{
\item{outcome}{a list object computed by \code{hier_functional}.}

\item{method}{\code{(method = 1)} diversity(alpha, gamma) based on Tsallis entropy (1988);\code{(method = 2)} beta diversity based on additive decomposition;
\code{(method = 3)} dissimilarity measure based on additive decomposition;\code{(method = 4)} diversity(alpha, gamma) based on Hill Number (1973);
\code{(method = 5)} beta diversity based on multiplicative  decomposition;\code{(method = 6)} dissimilarity measure based on multiplicative decomposition.}

\item{profile}{a selection of profile versus to diversity.q-profile \code{(profile = "q")};tau-profil \code{(prifile = "tau")}}
}
\description{
ggplot2 extension for outcome from \code{hier_functional}
}
\examples{
## Functional diversity
data(macro)
data(macro_mat)
data(macro_dis)
output3 = hier.functional(data = macro, mat = macro_mat, dis = macro_dis, q = c(0, 1, 2), FDtau = c(0.2, 0.4, 0.6, 0.8, 1))
gghier_functional(output3, method = 1, profile = "q")
output4 = hier.functional(data = macro, mat = macro_mat, dis = macro_dis, q = c(0, 1, 2), FDtype = "AUC")
gghier_functional(output4, method = 1, profile = "q")
}
