% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mainfunction.R
\name{gghier_taxonomy}
\alias{gghier_taxonomy}
\title{ggplot2 extension for outcome from \code{hier_taxonomy}}
\usage{
gghier_taxonomy(outcome, method = 1)
}
\arguments{
\item{outcome}{a list object computed by \code{hier_taxonomy}.}

\item{method}{\code{(method = 1)} diversity(alpha, gamma) based on Tsallis entropy (1988);\code{(method = 2)} beta diversity based on additive decomposition;
\code{(method = 3)} dissimilarity measure based on additive decomposition;\code{(method = 4)} diversity(alpha, gamma) based on Hill Number (1973);
\code{(method = 5)} beta diversity based on multiplicative  decomposition;\code{(method = 6)} dissimilarity measure based on multiplicative decomposition.}
}
\description{
ggplot2 extension for outcome from \code{hier_taxonomy}
}
\examples{
## Taxonomy diversity
data(macro)
data(macro_mat)
output1 = hier.taxonomy(data = macro, mat = macro_mat, q = seq(0, 2, 0.2))
gghier_taxonomy(output1, method = 1)
}
