% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mainfunction.R
\name{hier.phylogeny}
\alias{hier.phylogeny}
\title{Decomposition of phylogeny diversity}
\usage{
hier.phylogeny(
  data,
  mat,
  tree,
  q = seq(0, 2, 0.2),
  weight = "size",
  nboot = 20,
  conf = 0.95,
  type = "mle",
  decomposition = "relative"
)
}
\arguments{
\item{data}{\code{data.frames}}

\item{mat}{hierarchical structure of data.}

\item{tree}{a phylogenetic tree in Newick format for all observed species in the pooled assemblage.}

\item{q}{a numerical vector specifying the diversity orders. Default is \code{seq(0, 2, 0.2)}.}

\item{weight}{weight for relative decomposition.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing
sampling uncertainty and constructing confidence intervals. Bootstrap replications are generally time consuming. Enter 0 to skip the bootstrap procedures. Default is \code{20}.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is \code{0.95}.}

\item{type}{estimate type: estimate \code{(type = "est")}, empirical estimate \code{(type = "mle")}.Default is \code{"mle"}.}

\item{decomposition}{Relative decomposition: \code{(decomposition = "relative")}, Absolute decomposition: \code{(decomposition = "absolute")}.}
}
\description{
Decomposition of phylogeny diversity
}
\examples{
## Phylogeny diversity
data(macro)
data(macro_mat)
data(macro_tree)
output2 = hier.phylogeny(data = macro, mat = macro_mat, tree = macro_tree, q = seq(0, 2, 0.2))
output2
}
