% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mainfunction.R
\name{hier.taxonomy}
\alias{hier.taxonomy}
\title{Decomposition of taxonomy diversity}
\usage{
hier.taxonomy(
  data,
  mat,
  q = seq(0, 2, 0.2),
  weight = "size",
  nboot = 20,
  conf = 0.95,
  type = "mle",
  datatype = "abundance",
  decomposition = "relative"
)
}
\arguments{
\item{data}{\code{data.frames}}

\item{mat}{hierarchical structure of data.}

\item{q}{a numerical vector specifying the diversity orders. Default is \code{seq(0, 2, 0.2)}.}

\item{weight}{weight for relative decomposition.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing
sampling uncertainty and constructing confidence intervals. Bootstrap replications are generally time consuming. Enter 0 to skip the bootstrap procedures. Default is \code{20}.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is \code{0.95}.}

\item{type}{estimate type: estimate \code{(type = "est")}, empirical estimate \code{(type = "mle")}.Default is \code{"mle"}.}

\item{datatype}{data type of input data: individual-based abundance data \code{(datatype = "abundance")}, sampling-unit-based incidence frequencies data \code{(datatype = "incidence")}.}

\item{decomposition}{Relative decomposition: \code{(decomposition = "relative")}, Absolute decomposition: \code{(decomposition = "absolute")}.}
}
\description{
Decomposition of taxonomy diversity
}
\examples{
## Taxonomic diversity
data(macro)
data(macro_mat)
output1 = hier.taxonomy(data = macro, mat = macro_mat, q = seq(0, 2, 0.2))
output1
}
